% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_spectrum.R
\name{afni.spectrum}
\alias{afni.spectrum}
\title{AFNI color spectrum -> R object}
\usage{
afni.spectrum(coloreddata, img = NULL)
}
\arguments{
\item{coloreddata}{vector of data the spectrum has been applied to (can be just a range)}

\item{img}{the jpeg exported afni spectrum. Right click 'OLay' -> export ppm}
}
\description{
AFNI color spectrum -> R object
}
\examples{


}

