% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_spectrum.R
\name{afni_save_spectrum}
\alias{afni_save_spectrum}
\title{wrapper for optionally creating a pdf of plot_colorspectrum(afni.spectrum(...)))}
\usage{
afni_save_spectrum(volumemax, savename = NA, threshold = NA, ispos = F,
  img = NULL, ...)
}
\arguments{
\item{volumemax}{- maximum value for spectrum}

\item{threshold}{- where to cut off data}

\item{ispos}{- only postiive values?}

\item{...}{other options for plot_colorspectrum (lab="val", side=2, ax=T)}

\item{pdf}{name}
}
\description{
wrapper for optionally creating a pdf of plot_colorspectrum(afni.spectrum(...)))
}
\examples{
 # minmax <- nii_range('myfile.nii.gz[0]')
 minmax <- c(5,10)
 afni_save_spectrum(minmax[1],threshold=minmax[0],ispos=T,lab="F",ax=F)
 # thresholded w/neg postive and better axis labels
 afni_save_spectrum(10,threshold=5,ispos=F,lab="F",ax=F) 
 axis(side=2,at=c(-10,-5,5,10),labels=c("-10","-5","5","10"),las=2)
 # with image saved from afni color bar
 afni_save_spectrum(10,img="saved_spectrum.png") 
}
