% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_animated.R
\name{age_animate}
\alias{age_animate}
\title{age_animate}
\usage{
age_animate(d, weight_sigma = 2, rep_steps = 100, test_age = NULL, ...)
}
\arguments{
\item{d}{dataframe with age}

\item{weight_sigma}{how fast points disappear (higher is faster)}

\item{rep_steps}{how many steps to plot}

\item{test_age}{show only a single frame at 'test_age' instead of animating}

\item{...}{additional ggplot aesthetics. probably want x= and y=
#@importFrom viridis scale_color_viridis
#@importFrom gganimate ease_aes transition_time}
}
\description{
visualize age interaction with gganimate
}
\examples{
d <- data.frame(age=seq(10,30, length.out=8),
                x=runif(8,0,10),
                y=runif(8,0,10))
p  <- age_animate(d, rep_steps=16, x=x, y=y)
p2 <- age_animate(d, test_age=10) + aes(x=x,y=y) + cowplot::theme_cowplot() + theme(legend.position='none')
}
