% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_animated.R
\name{age_animate_weight}
\alias{age_animate_weight}
\title{age_animate_weight}
\usage{
age_animate_weight(d, frames_at, weight_sigma)
}
\arguments{
\item{d}{dataframe with age}

\item{frames_at}{vector of each "age" value to use as a frame}

\item{weight_sigma}{how fast points disappear (higher is faster)}
}
\value{
dataframe with length(frames_at) times more rows: and columns 'frameage' age@frame and 'w' weight for each point
}
\description{
duplicate dataframe for each element in frames_at
}
\examples{
d <- data.frame(age=seq(10,30, length.out=4))
movdata <- age_animate_weight(d, c(10,20,30), 2)
}
