% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_ungroup.R
\name{col_ungroup}
\alias{col_ungroup}
\title{col_ungroup: pull unique bits out of redudant column names to make dataframe longer
 e.g. Contact1Phone Contac2Phone -> Phone grpvar(=Conatct1,Contact)}
\usage{
col_ungroup(d, patt, grpvarcolname = "grpvar")
}
\arguments{
\item{d}{is the dataframe with redudant columns}

\item{patt}{is the pattern that matches redudant colnames (e.g. 'Contact[12]')}

\item{grpvarcolname}{is name to give to the new column (def. 'grpvar')}
}
\description{
col_ungroup: pull unique bits out of redudant column names to make dataframe longer
 e.g. Contact1Phone Contac2Phone -> Phone grpvar(=Conatct1,Contact)
}
\examples{
   # as prefix
   d2   <- data.frame(x=1:3,Contact1Phone=4:6,Contact2Phone=7:9,Contact1Address=2:4)
   d2.ug<- col_ungroup(d2,'Contact[12]')
   # with suffixe
   d1   <- data.frame(x=1:3,measure.a=4:6,measure.b=7:9,m2.a=2:4,m2.b=5:7)
   d1.ug<- col_ungroup(d1,'\\\\.(a|b)$')
   # watch out for patterns that match multiple parts of the same string
   # e.g. 'a|b' will be removed from 'measure.a' to make 'mesure'

}
