% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query.R
\name{db_query}
\alias{db_query}
\title{db_qeury -- query database}
\usage{
db_query(query, conn = lncd_pgconn())
}
\arguments{
\item{query}{is the sql string to run}

\item{conn}{is the optional connection. will be created from ~/.pgpass is left empty}
}
\description{
db_qeury -- query database
}
\examples{
 v   <- db_query('select * from visit limit 2')
 meg <- db_query(readr::read_file('/Volumes/Zeus/DB_SQL/queries/allMEG2016.sql'))
}
