% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthrate_gam.R
\name{find_covars_gam}
\alias{find_covars_gam}
\title{find_covars_gam}
\usage{
find_covars_gam(fml, ...)
}
\arguments{
\item{fml}{formula to get covariates from}

\item{...}{covariates to ignore (like agevar)}
}
\description{
pull out covariests from mgcv model formula
}
\examples{
 fml <- y ~ s(x) + x2 + s(x3, bs="re") + x4
 covars <- find_covars_gam(fml, x4) # x, x2
}
