% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthrate_gam.R
\name{gam_growthrate}
\alias{gam_growthrate}
\title{gam_growthrate}
\usage{
gam_growthrate(m, agevar, idvar = NULL, n.iterations = 10000,
  qnt = c(0.025, 0.975))
}
\arguments{
\item{m}{mgcv gam model object (only)}

\item{agevar}{variable for growth rate}

\item{idvar}{random effects/subject id variable, set to NULL if none}

\item{nnumber}{of iterations to run (quick)}

\item{qntquantiles}{to use for confidence interval}
}
\description{
posterior simulation for confidence intervals of local slope/growth rate (deriv) of mgcv gam modeled variable (age) - random effects: intercepts only (not predicted)
}
\details{
https://people.maths.bris.ac.uk/~sw15190/mgcv/tampere/mgcv-advanced.pdf
https://stats.stackexchange.com/questions/190348/can-i-use-bootstrapping-to-estimate-the-uncertainty-in-a-maximum-value-of-a-gam
}
\examples{
 d <- read.csv("/Volumes/Phillips/R03_BehavioralBTC/data/btc_R03scoredmeasures_20190313.csv") \%>\%
   group_by(id) \%>\%
   mutate(visit=rank(d8))
 f <- f1score ~ s(Ageatvisit) + s(visit) + s(id, bs="re")
 m <- mgcv::gam(f, data=d)
 ci <- gam_growthrate(m, 'Ageatvisit', 'id')
}
