% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthrate_gam.R
\name{gam_maturation_point}
\alias{gam_maturation_point}
\title{gam_maturation_point}
\usage{
gam_maturation_point(ci)
}
\arguments{
\item{ci}{growthrate_gam output (confidence interval and derivitive)}
}
\description{
get maturation point from confidence interval dataframe
}
