% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthrate_gam.R
\name{growthrate_gam}
\alias{growthrate_gam}
\title{growthrate_gam}
\usage{
growthrate_gam(m, agevar, n = 10000, qnt = c(0.025, 0.975))
}
\arguments{
\item{m}{-- gam model object}

\item{agevar}{-- variable for growthrate}

\item{n}{-- number of iterations to run (quick)}

\item{qnt}{-- quantiles to use for confidence interval}
}
\description{
confidence intervals change of gam modeled variable (age)
}
\details{
https://stats.stackexchange.com/questions/190348/can-i-use-bootstrapping-to-estimate-the-uncertainty-in-a-maximum-value-of-a-gam
}
\examples{
 d <- read.csv("/Volumes/Phillips/R03_BehavioralBTC/data/btc_R03scoredmeasures_20190313.csv") \%>\%
   group_by(id) \%>\%
   mutate(visit=rank(d8))
 f <- f1score ~ s(Ageatvisit) + s(visit) + s(id, bs="re")
 m <- gam(f, data=d)
 ci <- growthrate_gam(m, 'Ageatvisit')
}
