% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hera.R
\name{hera}
\alias{hera}
\title{Build OS (windows or unix) agnostic mount path to hera data share}
\usage{
hera(path, ...)
}
\arguments{
\item{path}{location on hera. use '/' as path separater}

\item{...}{additional path components. will concat with file.path}
}
\description{
Build OS (windows or unix) agnostic mount path to hera data share
}
\examples{
hera("/Volumes/Hera/scratch/")
hera("H:/scratch/")
hera("scratch/")
}
