% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICCforlongformat}
\alias{ICCforlongformat}
\title{ICClong}
\usage{
ICCforlongformat(d, vars, subjectvar, visitvar, maxvisits = 2)
}
\arguments{
\item{d}{long format (row per visit, possibly many rows per subject)}

\item{vars}{which columns to return ICC values for}

\item{subjectvar}{column uniquely identifying subject}

\item{visitvar}{column uniquely identifying visit}

\item{maxvisits}{cutoff for number of visits any one subject contributes}
}
\description{
ICC wrapper for wide format data frame
}
\examples{
d <- data.frame(id=rep(1:6, 3), visit=rep(1:3, each=6),
                a=runif(18,0,10),
                b=runif(18,0,10))
d_ICCs <- ICCforlongformat(d, c('a','b'), "id", "visit")
}
