% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_conv.R
\name{ijk.afni2oro}
\alias{ijk.afni2oro}
\title{ijk.afni2oro: given afni gui voxel dims, get MNI (LPI) dims for oro.nifiti::readNifTI's @.Data matrix}
\usage{
ijk.afni2oro(ijk, dm)
}
\arguments{
\item{ijk:}{i,j,k vector read from afni gui}

\item{dm:}{dim(x@.Data)}
}
\description{
ijk.afni2oro: given afni gui voxel dims, get MNI (LPI) dims for oro.nifiti::readNifTI's @.Data matrix
}
\examples{
\dontrun{
 x <- oro.nifti::readNIfTI('betas.nii.gz')
 idx <- ijk.afni2oro(c(25,20,30), dim(x) )
}
}
