% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_conv.R
\name{ijk.oro2afni}
\alias{ijk.oro2afni}
\title{ijk.oro2afni: given index of oro nifti @.Data matrix, get index to give to AFNI's "jump to ijk"}
\usage{
ijk.oro2afni(ijk, dm)
}
\arguments{
\item{ijk:}{i,j,k index of oro nifti matrix}

\item{dm:}{dim of oro nifti matrix [dim(x@.Data)]}
}
\description{
ijk.oro2afni: given index of oro nifti @.Data matrix, get index to give to AFNI's "jump to ijk"
}
\examples{
\dontrun{
 x <- oro.nifti::readNIfTI('betas.nii.gz')
 dm <- dim(x)
 mx <- arrayInd(which.max(d),dm)
 ijk.oro2afni(mx[1:3], dm )
}
}
