% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_label_match.R
\name{interactive_label_match}
\alias{interactive_label_match}
\title{interactive_label_match:  match labels from one string vector with another}
\usage{
interactive_label_match(labels1, labels2)
}
\arguments{
\item{labels1}{are the set of labels to match}

\item{labels2}{are where to pull potential matches from}
}
\description{
interactive_label_match:  match labels from one string vector with another
}
\examples{
   # as prefix
   labels1 <- ysr_datadict_nondirect$Field.Label
   label2 <- asr_datadict_nondirect$Field.Label
   interactive_label_match(labels1, labels2)
}
