% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query.R
\name{ld8_age}
\alias{ld8_age}
\title{ld8_age - merge input dataframe with ages calc from db}
\usage{
ld8_age(d, colname = "ld8", selectquery = "dob, sex")
}
\arguments{
\item{d}{-- dataframe with an lunaid_yyyymmdd column}

\item{colname}{-- column containing ld8}

\item{selectquery}{-- what columns to select from db}
}
\description{
ld8_age - merge input dataframe with ages calc from db
}
\examples{
 rist  <- ld8_age(data.frame(ld8=c("11654_20180602","11671_20180723")),"ld8")
}
