% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld8.R
\name{ld8from}
\alias{ld8from}
\title{Extract lunaid_date from text}
\usage{
ld8from(s)
}
\arguments{
\item{s}{(vector of) string to extract luna_yyyymmdd}
}
\description{
Extract lunaid_date from text
}
\examples{
ld8from("/path/to/12345_20191011/file.txt")
}
