% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tools.R
\name{lm_list}
\alias{lm_list}
\title{lm_list}
\usage{
lm_list(d, formula, ...)
}
\arguments{
\item{d}{input long format data.frame w/grouping columns specified later}

\item{formula}{lm formula to model each split}

\item{...}{columns in d to split dataframe (unique groups)}
}
\description{
build a list of lm models from group columns
}
\examples{
 mpg_cyl_mdls <- lm_list(mtcars, mpg~cyl, "gear", "am")
 mpg_cyl_mdls <- lm_list(mtcars, mpg~cyl, c("gear", "am")) # same as above
}
