% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tools.R
\name{lm_pval}
\alias{lm_pval}
\title{lm_pval}
\usage{
lm_pval(m, coef_row = 2, pr_name = "Pr(>|t|)")
}
\arguments{
\item{m}{lm model}

\item{coef_row}{what row to extract pvalue from (def=2, likely slope)}

\item{pr_name}{column name for pvalue. default="Pr(>|t|)"}
}
\description{
extract single pvalue from lm model. shortcut for coef(summary(m))
}
\examples{
 lm_pval(lm(mpg~cyl, mtcars)) # 6.112687e-10
}
