% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_extract.R
\name{lmer_extract}
\alias{lmer_extract}
\title{lmer_extract
 get (t, chisq, p) values from a single variable in a model
 ONLY USE WHERE FACTOR has more than two levels -- chisq will report without any factor instead of without the specified factor}
\usage{
lmer_extract(model, varname, factorname = NULL)
}
\arguments{
\item{model}{is a lm model}

\item{varname}{is the variable of interest (string). ie. rowname in car::Anova to extract Chisq and Pr>Chisq}

\item{factorname}{is optional. factor of var (string) as named by summary(model), e.g. 'Female01female'.  ie. rowname in summary to extract tvalue}
}
\description{
lmer_extract
 get (t, chisq, p) values from a single variable in a model
 ONLY USE WHERE FACTOR has more than two levels -- chisq will report without any factor instead of without the specified factor
}
\examples{
  m <- lme4::lmer(uptake ~ conc + Type + (1 | Treatment), CO2)
  lmer_extract(m, "conc")
  lmer_extract(m, "Type", "TypeMississippi")
}
