% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query.R
\name{lncd_pgconn}
\alias{lncd_pgconn}
\title{lncd_pgconn -- create connection to lncd postgres database}
\usage{
lncd_pgconn(info = pgpassread())
}
\arguments{
\item{info}{is a named list with db, host, user, and pass. will be created from ~/.pgpass is left empty}
}
\description{
lncd_pgconn -- create connection to lncd postgres database
}
\examples{
conn <- lncd_pgconn()
# lapply(dbListConnections(RPostgreSQL::PostgreSQL()),dbDisconnect)
}
