% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunaize.R
\name{lunaize}
\alias{lunaize}
\title{Re-theme plot a la Luna
This function stylizes a ggplot2 plot to match the aesthetic of Dr. Luna
 * Removes grid lines and borders
 * Increases font size 
 * Increases gap between x and y axis titles (by adding "\n" to the labels)}
\usage{
lunaize(p, ajust = 0.05)
}
\arguments{
\item{p}{plot to retheme}

\item{ajust}{vjust and hjust values for axis text Defauts to .05}
}
\description{
Re-theme plot a la Luna
This function stylizes a ggplot2 plot to match the aesthetic of Dr. Luna
 * Removes grid lines and borders
 * Increases font size 
 * Increases gap between x and y axis titles (by adding "\n" to the labels)
}
\examples{
p <- ggplot(iris,aes(x=Sepal.Length,y=Petal.Width))+geom_point()
lunaize(p) # or with dplyr: p \%>\% lunaize
}

