% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vox_cor.R
\name{mask_nii}
\alias{mask_nii}
\title{mask_nii:  mask a nii by a mask, returns num matrix w/ rows of time, columns of i_j_k voxels}
\usage{
mask_nii(fname_or_oro, mask)
}
\arguments{
\item{fname_or_oro}{a nifti filename or oro.nifti object}

\item{mask}{a "logi" booleen matrix (likely from \code{\link{read_mask}})}
}
\description{
mask_nii:  mask a nii by a mask, returns num matrix w/ rows of time, columns of i_j_k voxels
}
\examples{
   \dontrun{
    mask <- read_mask("striatum_mask.nii.gz") # 1827 vx in mask
    data <- mask_nii("subj_ts.nii.gz",mask)  # 320 trs
    # data:     num [1:320, 1:1827] 
    # colnames: chr [1:1827] "47_61_30" ...
   }
}
