% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion.R
\name{motion_eangle}
\alias{motion_eangle}
\title{mean Euler Angle
see http://dx.doi.org/10.1016/j.neuroimage.2011.07.044}
\usage{
motion_eangle(d)
}
\arguments{
\item{x}{- motion dataframe w/first 3 columsn are rots or motion.par file}
}
\description{
mean Euler Angle
see http://dx.doi.org/10.1016/j.neuroimage.2011.07.044
}
