% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion.R
\name{motion_nfdabove}
\alias{motion_nfdabove}
\title{number fd > treshold (def .1mm)}
\usage{
motion_nfdabove(x, thres = 0.1)
}
\arguments{
\item{x}{- vector or fd file}

\item{thres}{- count values above this}
}
\description{
number fd > treshold (def .1mm)
}
