% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion.R
\name{motion_readpar}
\alias{motion_readpar}
\title{extract motion paramiters ordered like mcflirt}
\usage{
motion_readpar(f)
}
\arguments{
\item{f}{- 'motion.par' file}
}
\description{
extract motion paramiters ordered like mcflirt
}
