% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roitempcorr_wide.R
\name{parseROItempcor}
\alias{parseROItempcor}
\title{parseROItempcor - read adj_pearson.txt and diagnostic_files.txt into wide}
\usage{
parseROItempcor(adj_glob, roi_names = NULL, idfunc = LNCDR::ld8from)
}
\arguments{
\item{adj_glob}{file path glob to adj_pearson.txt fles}

\item{roi_names}{names to use instead of searcing for diagnostic_files.txt and using first column}

\item{idfunc}{default ld8from. function to pull id from filename}
}
\description{
parseROItempcor - read adj_pearson.txt and diagnostic_files.txt into wide
}
\examples{
adjfilesglob<-"/Volumes/Hera/Projects/BTC_Diss/rs_20200321/striatalseeddata/*_20161*/frontostrialROItempcor20210106.rac1.adj_pearson.txt"
rac1 <- parseROItempcor(adjfilesglob)
}
