% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_spectrum.R
\name{plot.colorspectrum}
\alias{plot.colorspectrum}
\title{plot a color spectrum read in by afni.spectrum}
\usage{
plot.colorspectrum(colorval)
}
\arguments{
\item{colorval}{the list output of afni.spectrum}
}
\description{
plot a color spectrum read in by afni.spectrum
}
\examples{
 colorval <- afni.spectrum(-5:5)
 plot.colorspectrum() 
}

