% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_spectrum.R
\name{plot_colorspectrum}
\alias{plot_colorspectrum}
\title{plot a color spectrum read in by afni.spectrum}
\usage{
plot_colorspectrum(colorval, lab = "val", side = 2, ax = T, thres = NA)
}
\arguments{
\item{colorval}{the list output of afni.spectrum}

\item{lab}{label for the relevant axis, defaults to useless 'val'}

\item{side}{which direction to draw the graph 1 horz, 2 vertical. matches par's meaning of side}

\item{ax}{T/F weither or not this function should draw an axis. Default T;label for max,middle, and min}
}
\description{
plot a color spectrum read in by afni.spectrum
}
\examples{
 # Simple
 colorval <- afni.spectrum(5)
 plot_colorspectrum(colorval,'F') 

 # Threshold, provide on axis
 # limit to a threshold
 plot_colorspectrum(colorval,'F',thres=2.68) 
 # only  positive
 cv <- afni.spectrum(c(0,5))
 plot_colorspectrum(cv,'F') 
 # start at threshold
 cv <- cv[cv$invals>2.68,]
 plot_colorspectrum(cv,'F') 
 # make text 150\% of the normal size for the labels
 par(cex.lab=1.5)
 # plot
 plot_colorspectrum(cv,'F',ax=F) 
 # add our own axis
 axis(side=2,at=c(3,4,5),labels=c("3","","5"),las=2)
}
