% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{pubmed_search}
\alias{pubmed_search}
\title{Pubmed search into a dataframe}
\usage{
pubmed_search(query, prefix)
}
\arguments{
\item{query}{-- a pubmed search query}

\item{prefix}{-- xml file prefix: what to search/call temporary output}
}
\value{
dataframe with columns title, authors, year, journal, abstract, doi
}
\description{
From pubmed query to dataframe
}
\examples{
  sub_use <- pubmed_search('"substance use" risk nueroimaging', 'tmp/fmri_substance')
}
