% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qassertm.R
\name{qassert_all}
\alias{qassert_all}
\title{Check single rule against all or enumerated variables}
\usage{
qassert_all(rule, ...)
}
\arguments{
\item{rule}{[[\code{character}]]\cr
Any number of named arguments like var.name=rule\cr
See details of \code{\link{qtest}} for rule explanation.}

\item{...}{[[\code{var.name}]]\cr
list of unquoted variables that should match \code{rule}
if none provided, uses all variables in local scope}
}
\value{
See \code{\link{qassert}}.
}
\description{
Ergonomic wraper for \code{\link{qassert}} to check multiple variables against a single rule
}
\examples{
x <- TRUE; y <- FALSE
# check all variables, scoped inside function
foofunc <- function(x, y) { qassert_all('n'); print(x); }
foofunc(1, 2)
# only enumerated variables
qassert_all('l', x, y)

}
\seealso{
\code{\link{qassertm}}
}
