% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qassertm.R
\name{qassertm}
\alias{qassertm}
\title{Quick checks for multiple variable=rule enumerated as arguments}
\usage{
qassertm(...)
}
\arguments{
\item{...}{[\code{var.name}=[\code{character}]]\cr
Any number of named arguments like var.name=rule\cr
See details of \code{\link{qtest}} for rule explanation.}
}
\value{
See \code{\link{qassert}}.
}
\description{
Ergonomic wraper for \code{\link{qassert}} to check multiple variable=rule pairs
}
\examples{
x=1:10; y=TRUE
qassertm(x="i+", y="b")

with(data.frame(a=1:26,b=letters), qassertm(a="i+", b="s"))
}
\seealso{
\code{\link{qtest}}, \code{\link{qassert}}
}
