% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_auth}
\alias{qualtrics_auth}
\title{authenticate with qualtrics using api token in config file}
\usage{
qualtrics_auth(config = "~/.qualtrics.ini")
}
\arguments{
\item{config}{- where to read qualtrics settings}
}
\description{
default config expected in $HOME/.qualtrics.ini
  [api]
  api_token= 67dfRZ2ibBRPxbLmbxxxxxxxxxxxxxxxxxxxxxxx
  root_url = https://pitt.co1.qualtrics.com
}
