% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_battery_names}
\alias{qualtrics_battery_names}
\title{separate survey names into}
\usage{
qualtrics_battery_names(srvy)
}
\arguments{
\item{srvy}{- dataframe like qualtRics::fetch_survey output (e.g. qualtircs_surveys()[[1]])}
}
\description{
expect names like SRVY_X.Q1 ... SRVY_X.Q10a, SRVY_Y.Q1 ... SRVY_Y.Q20
 this is set in design of qualtrics survey (Q title not seen by participants)
 included here for reference
}
