% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_labels}
\alias{qualtrics_labels}
\title{extract label from each column of a dataframe}
\usage{
qualtrics_labels(d)
}
\arguments{
\item{d}{- dataframe like qualtRics::fetch_survey output (e.g. qualtircs_surveys()[[1]])}
}
\description{
extract label from each column of a dataframe
}
