% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_surveys}
\alias{qualtrics_surveys}
\title{pull surveys matching a pattern into a list of dataframes}
\usage{
qualtrics_surveys(
  name_pattern = ".*",
  active = T,
  config = "~/.qualtrics.ini",
  cautious_infer = FALSE
)
}
\arguments{
\item{name_pattern}{- grep pattern to match against survey names (e.g. 'Covid')}

\item{active}{- include only active surveys (default = T)}

\item{config}{- where to read qualtrics settings}

\item{cautious_infer}{- tryCatch qualtRics:::infer_data_types. lose metadata but read in when there are errors. NULL for no covert}
}
\description{
read all surveys from qualtrics accessible to api key provided in config
 see help(qualtircs_auth) for more on config format
 consider combining returned list with data.table::rbindlist or dplyr::bind_rows
}
