% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vox_cor.R
\name{read_mask}
\alias{read_mask}
\title{read_mask: read a nifti file as a mask, return "logi" (boolean) array}
\usage{
read_mask(fname)
}
\arguments{
\item{fname}{file path to nifti file}
}
\description{
read_mask: read a nifti file as a mask, return "logi" (boolean) array
}
\examples{
   \dontrun{
    seed <- read_mask("striatum_mask.nii.gz")
    str(seed) # logi [1:84, 1:100, 1:84]
   }
}
