% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_data_sim.R
\name{retentionadjustvisits}
\alias{retentionadjustvisits}
\title{retentionadjustvisits}
\usage{
retentionadjustvisits(nsubject, nvisits, retention_or_total)
}
\arguments{
\item{nsubject}{how many subjects}

\item{nvisits}{how many visits each}

\item{retention_or_total}{ratio of retained (0<x<1) or total (x>=1)}
}
\description{
simulate multilevel bivariate normal dist with a random intercept
}
\examples{
retentionadjustvisits(100,2,.8)  # using retention
retentionadjustvisits(100,2,160) # using total
}
