% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roitempcorr_wide.R
\name{roicormat_wide}
\alias{roicormat_wide}
\title{roicormat_wide - extract upper tri as single vector. name with rois}
\usage{
roicormat_wide(input_file_or_mat, roi_names = NULL)
}
\arguments{
\item{input_file_or_mat}{- dataframe or path to e.g. adj_pearson.txt}

\item{roi_names}{- optional names for rows and columns (symmetric)}
}
\description{
roicormat_wide - extract upper tri as single vector. name with rois
}
