% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save1D.R
\name{save1D}
\alias{save1D}
\title{Create .1D files froma data frame
 * works on a dataframe with 'block' column
 * cats a 1D file, optionally sink'ed to a specified file}
\usage{
save1D(
  d,
  colname = 1,
  fname = NULL,
  dur = NULL,
  amp = NULL,
  nblocks = NULL,
  createdirs = FALSE
)
}
\arguments{
\item{d}{dataframe with manditory 'block' column}

\item{colname}{is the column to use for onset time}

\item{fname}{is where to save the output 1D. if NULL does not save to a file}

\item{dur}{if not null, what to set as the duration part of onset:duration}

\item{amp}{if not null, what to set as the amplitude part of onset*amplitude}

\item{nblocks}{expected number of blocks or runs. if null will use max}
}
\description{
Create .1D files froma data frame
 * works on a dataframe with 'block' column
 * cats a 1D file, optionally sink'ed to a specified file
}
\examples{
save1D( data.frame(onset=1:9,durcol=1,block=1:3), 'onset',dur='durcol') 
}
