% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_data_sim.R
\name{sim_multilevel_bivariatenormal_randintercept}
\alias{sim_multilevel_bivariatenormal_randintercept}
\title{sim_multilevel_bivariatenormal_randintercept}
\usage{
sim_multilevel_bivariatenormal_randintercept(nsubject, nvisits,
  interceptcor, ICCs)
}
\arguments{
\item{nsubject}{how many subjects}

\item{nvisits}{how many visits each}

\item{interceptcor}{intercept correlation}

\item{ICCs}{inter class correlation}
}
\description{
simulate multilevel bivariate normal dist with a random intercept
}
\examples{
  sim_multilevel_bivariatenormal_randintercept(100,2,.3,.4)
}
