% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_data_sim.R
\name{sim_multilevel_bivariatenormal_randinterceptandslope}
\alias{sim_multilevel_bivariatenormal_randinterceptandslope}
\title{sim_multilevel_bivariatenormal_randinterceptandslope}
\usage{
sim_multilevel_bivariatenormal_randinterceptandslope(nsubject, nvisits,
  interceptcor, slopecor, slopeintcorbetween, slopeintcorwithin, ICCs)
}
\arguments{
\item{nsubject}{how many subjects}

\item{nvisits}{how many visits each}

\item{interceptcor}{intercept correlation}

\item{slopecor}{slope correlation}

\item{slopeintcorbetween}{slope between subject correlation}

\item{slopeintcorwithin}{slope within subject correlation}

\item{ICCs}{inter class correlation}
}
\description{
simulate multilevel bivariate normal dist with a random intercept AND SLOPE
}
\examples{
 sim_multilevel_bivariatenormal_randinterceptandslope(100,2,.1,.5,.6,.3,.4)
}
