% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query.R
\name{task_query}
\alias{task_query}
\title{task_query - return all rows for a given task from psql database}
\usage{
task_query(task, columns = "measures")
}
\arguments{
\item{task}{-- like "DrugUse", "Demographics", etc}

\item{columns}{-- sql list of columns (default "measures")}
}
\description{
task_query - return all rows for a given task from psql database
}
\examples{
 rist  <- task_query("RIST", "measures->'rist_ristindex' as ristindex")
 demo  <- task_query("Demographics")
}
