% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick_seq.R
\name{tick_seq}
\alias{tick_seq}
\title{generate a sequence and minor/major tick labs for a ggplot axis}
\usage{
tick_seq(st, ed, by, labels.at)
}
\arguments{
\item{st}{start of sequence}

\item{ed}{end of sequence}

\item{by}{interval steps}

\item{labels.at}{value-of/where-to-place labels}
}
\description{
generate a sequence and minor/major tick labs for a ggplot axis
}
\examples{
 ts <- tick_seq(-10,10,5, c(-10,10,0) )
 # RETURNS
 #  ts$labels "-10" ""  "0"   ""  "10" 
 #  ts$breaks  -10  -5   0    5   10
 p <- ggplot()
 p + scale_x_continuous(breaks=ts$breaks,labels=ts$labels)
}

