% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query.R
\name{unnestjson}
\alias{unnestjson}
\title{unnest json(b) object}
\usage{
unnestjson(d, column = "measures")
}
\arguments{
\item{d}{the dataframe contianing a column to unnest}

\item{column}{the column (string, in quotes) containing the json string to be unnested. defaults to 'measures'}
}
\description{
unnest json(b) object
}
\examples{
 v  <- db_query('select * from visit_tasks where task like 'Sen\%' limit 2') \%>\% unnestjson
}
