% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppsp_score.R
\name{uppsp_scoring}
\alias{uppsp_scoring}
\title{upps_scoring: score upps}
\usage{
uppsp_scoring(uppspdf)
}
\arguments{
\item{uppspdf}{- dataframe of recorded uppsp responses}
}
\description{
upps_scoring: score upps
}
