% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vox_cor.R
\name{vox_cor}
\alias{vox_cor}
\title{vox_cor: voxelwise correlation for timeseries between two masks}
\usage{
vox_cor(tsnii, seed, trgmsk)
}
\arguments{
\item{tsnii}{4d timeseries as file name, oronifti object, or 4d timeseries matrix}

\item{seed}{1d numeric matrix or "logi" boolean matrix (seed_mask@.Data==1, see \code{\link{read_mask}})}

\item{trgmsk}{"logi" boolean matrix (trg_mask@.Data==1, see \code{\link{read_mask}})}
}
\description{
vox_cor: voxelwise correlation for timeseries between two masks
}
\examples{
   # usage
   \dontrun{
   seed <- read_mask("striatum_mask.nii.gz")
   target <- read_mask("gm_mask.nii.gz")
   target <- target & ! seed
   allcors <- vox_cor("subj_ts.nii.gz",seed,target)
   # passing seed as a vector
   allcors_mean <- vox_cor("subj_ts.nii.gz",apply(seed,2,mean),target)
   }
}
