% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall_group}
\alias{waterfall_group}
\title{waterfall plot data}
\usage{
waterfall_group(d)
}
\arguments{
\item{d}{dataframe with columns id (repeated) and  age}
}
\description{
create age_id ranked id for data grouped by start age
}
\examples{
 d <- data.frame(age=c(10,20,25, 13,14), id=c(100,100,100, 200,200))
 d_grouped <- waterfall_group(d)
}
