% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorizesd.R
\name{winsorize_sd}
\alias{winsorize_sd}
\title{winsozrize_sd - set above/below devation range to range max/min}
\usage{
winsorize_sd(x, devno, dev = "sd")
}
\arguments{
\item{x}{vector to be winsorized}

\item{devno}{number of deviations to winsorize}

\item{dev}{"sd" (standard deviation) or "mad" (median absolute deviation)}
}
\description{
winsozrize_sd - set above/below devation range to range max/min
}
