% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_date.R
\name{xlsx_date}
\alias{xlsx_date}
\title{number to date for vector/column from excel sheet with mixed value types}
\usage{
xlsx_date(x, msg = NULL, minyear = 2000, maxyear = year_() + 2)
}
\arguments{
\item{x}{(vector of) would be dates to transform back}
}
\description{
readxl::read_excel() on a column with mixed formated cells will not easily parse dates
consider read_excel(..., col_types = c("date","guess",...))
unless column is mixed valued (e.g. "42068", "1/31/2010,1/2/2011", ...)
NB. excel's date origin is 1899-12-30
}
\examples{
print(xlsx_date(c(41977,"42068"))) # "2014-12-04" "2015-03-05"
xlsx_date("10977") # error out of range (1933-01-14)
}
