% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{zscore: (vector - mean)/sd}
\usage{
zscore(x)
}
\arguments{
\item{x}{is the numeric vector to zscore}
}
\description{
zscore: (vector - mean)/sd
}
\examples{
   z<-zscore(1:10)
   mean(z) == 0
}
