% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscorecols}
\alias{zscorecols}
\title{zscores all columns in a df except those whose names are passed as "ignorecolnames"}
\usage{
zscorecols(df, ignorecolnames, VERBOSE = T)
}
\arguments{
\item{df}{is dataframe which columns to zscore}

\item{ignorecolnames}{is column vector of strings with column names to ignore}

\item{VERBOSE}{flag for printing factor during loop iteration}
}
\description{
zscores all columns in a df except those whose names are passed as "ignorecolnames"
}
\examples{
  d <- data.frame(a=1:10, b=seq(10,100,by=10), c=rep('a',10) )
  LNCDR::zscorecols( d, 'c' )
}
