% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combatwrap.R
\name{combatwrap}
\alias{combatwrap}
\title{haromize: wrap around neuroCombat}
\usage{
combatwrap(d, combatcols, batchvar, covars, eb)
}
\arguments{
\item{d}{dataframe}

\item{combatcols}{columns to use from d}

\item{batchvar}{columns TODO: describe}

\item{covars}{columns TODO: describe}

\item{eb}{TODO: describe}
}
\description{
haromize: wrap around neuroCombat
}
\examples{
# TODO: provide example
}
