% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_match.R
\name{date_match}
\alias{date_match}
\title{match dates in a data frame by one-to-many merge and rank. Useful to match behavioral to brain over multiple timepoints.}
\usage{
date_match(d1, d2, idcol, datecol1, datecol2 = datecol1, maxdatediff = 180,
  maxrank = 2 - 0.001)
}
\arguments{
\item{d1}{is the main dataframe}

\item{d2}{is the supplimental dataframe with dates to match}

\item{datecol1}{is date column one, this will be the id column}

\item{maxdatediff}{is the maximum time between matches (default 180 days)}

\item{maxrank}{is highest index value of datediff ranked. set to 1 to ingore ties (deafult is 1.999)}

\item{datecal2}{is date column two. this will be the many in one-to-many merge. defaults to same name as datacol1}
}
\description{
match dates in a data frame by one-to-many merge and rank. Useful to match behavioral to brain over multiple timepoints.
}
\examples{
  dummy1 <- data.frame(id=c(1,2,3,3),d=c(20120302,20120302,20120302,20150302),j=1)
  dummy2 <- data.frame(id=c(1,2,3,3,3),d=c(20120202,20120402,20120902,20130902,20150102),j2=2)
  dummy1$d <- lubridate::ymd(dummy1$d)
  dummy2$d <- lubridate::ymd(dummy2$d)
  date_match(dummy1,dummy2,'id','d')


}
