% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthrate_gam.R
\name{gam_growthrate_plot}
\alias{gam_growthrate_plot}
\title{plot gam factor with deriv}
\usage{
gam_growthrate_plot(d, model, ci, agevar, yvar, idvar,
  plotsavename = NULL, xplotname = "Age", yplotname = yvar)
}
\arguments{
\item{d}{dataframe model was built on (for actual points)}

\item{model}{gam model (for predicted line)}

\item{ci}{growthrate_gam output (confidence interval and derivitive)}

\item{agevar}{column name of age var e.g. 'Ageatvisit'}

\item{yvar}{model yvar e.g. 'f1score'}

\item{idvar}{line grouping var e.g., 'lunaid'}

\item{plotsavename}{pdf output name e.g. 'growth.pdf', not saved when NULL}

\item{xplotname}{'Age'}

\item{yplotname}{'f1score'}
}
\description{
plot output of growthrate_gam
}
\examples{
 
 m <- gam(f1score ~ s(Ageatvisit) + s(visit) + s(id, bs="re"), data=d)
 ci <- growthrate_gam(m, 'Ageatvisit')
plotgammfactorwithderiv(d, m, ci, 'Ageatvisit','f1score','id')
}
