% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthrate_gam.R
\name{gam_growthrate_plot}
\alias{gam_growthrate_plot}
\title{plot gam factor with deriv}
\usage{
gam_growthrate_plot(d, model, ci, agevar, idvar = NULL,
  yvar = as.character(model$formula[2]), plotsavename = NULL,
  xplotname = "Age", yplotname = yvar, draw_maturation = T,
  draw_points = T)
}
\arguments{
\item{d}{dataframe model was built on (for actual points)}

\item{model}{gam model (for predicted line)}

\item{ci}{growthrate_gam output (confidence interval and derivitive)}

\item{agevar}{column name of age var e.g. 'Ageatvisit'}

\item{idvar}{line grouping var e.g., 'lunaid', set to NULL if no random effect in model}

\item{yvar}{model yvar e.g. 'f1score', default pulled from model formula}

\item{plotsavename}{PDF output name e.g. 'growth.pdf', not saved when NULL}

\item{xplotname}{'Age'}

\item{yplotname}{'f1score', default is yvar (model yvar)}

\item{draw_maturation}{T|F, show dotted line on first maturation point}

\item{draw_points}{T|F, show individual points as scatter plot over gam fit line}
}
\description{
plot output of growthrate_gam
}
\examples{

 m <- mgcv::gam(f1score ~ s(Ageatvisit) + s(visit) + s(id, bs="re"), data=d)
 ci <- gam_growthrate(m, 'Ageatvisit')
 gam_growthrate_plot(d, m, ci, 'Ageatvisit', 'id')

 # need to explicity set id to NULL if no random effect
 m <- mgcv::gam(f1score ~ s(Ageatvisit), data=d)
 ci <- gam_growthrate(m, 'Ageatvisit', idvar=NULL)
 gam_growthrate_plot(d, m, ci, 'Ageatvisit', idvar=NULL)
}
