% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_label_match.R
\name{interactive_label_match}
\alias{interactive_label_match}
\title{interactive_label_match:  match labels from one string vector with another}
\usage{
interactive_label_match(
  labels1,
  labels2,
  accept_single = FALSE,
  prompt_equal = FALSE,
  diffprint = TRUE
)
}
\arguments{
\item{labels1}{are the set of labels to match}

\item{labels2}{are where to pull potential matches from}

\item{accept_single}{default FALSE. if only one match, accept it.}

\item{prompt_equal}{default FALSE. set to true if you want to confirm all matches}

\item{diffprint}{default TRUE. show colored diff. only useful in terminal}
}
\description{
interactive_label_match:  match labels from one string vector with another
}
\examples{
   # as prefix
   labels1 <- ysr_datadict_nondirect$Field.Label
   label2 <- asr_datadict_nondirect$Field.Label
   interactive_label_match(labels1, labels2)
}
