% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_extract.R
\name{lmer_extract}
\alias{lmer_extract}
\title{get pvalues from a model}
\usage{
lmer_extract(model, var, factorname = NULL)
}
\arguments{
\item{model}{is a lm model}

\item{var}{is the variable of interest (string)}

\item{factorname}{is optional factor of var}
}
\description{
get pvalues from a model
}
\examples{
m <- lm(model=
lmer_extract(m
}
