% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_spectrum.R
\name{plot.colorspectrum}
\alias{plot.colorspectrum}
\title{plot a color spectrum read in by afni.spectrum}
\usage{
plot.colorspectrum(colorval, lab = "val", side = 2)
}
\arguments{
\item{colorval}{the list output of afni.spectrum}

\item{lab}{label for the relevant axis, defaults to useless 'val'}

\item{side}{which direction to draw the graph 1 horz, 2 vertical. matches par's meaning of side}
}
\description{
plot a color spectrum read in by afni.spectrum
}
\examples{
 colorval <- afni.spectrum(-5:5)
 #svg('colorbar.svg')
 par(cex.lab=1.5) # 150\% of the normal size for the labels
 plot.colorspectrum(colorval) 
 #dev.off()
}

