% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_surveys}
\alias{qualtrics_surveys}
\title{pull surveys matching a pattern into a list of dataframes}
\usage{
qualtrics_surveys(name_pattern, active = T, config = "qualtrics.ini")
}
\arguments{
\item{name_pattern}{- grep pattern to match against survey names (e.g. 'Covid')}

\item{active}{- include only active surveys (default = T)}

\item{config}{- where to read qualtrics settings}
}
\description{
pull surveys matching a pattern into a list of dataframes
}
