% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppsp_score.R
\name{uppsp_scoring}
\alias{uppsp_scoring}
\title{Author: BTC
upps_scoring - scores uppsp 59-item version
  * wide format data only
  * items must start with column number
  * returns five factor scores and total}
\usage{
uppsp_scoring(uppspdf)
}
\arguments{
\item{uppspdf}{- dataframe of recorded uppsp responses}
}
\description{
Author: BTC
upps_scoring - scores uppsp 59-item version
  * wide format data only
  * items must start with column number
  * returns five factor scores and total
}
\examples{
  uppsp<-readxl::read_excel("/Volumes/Phillips/mMR_PETDA/scripts/txt/PET_Sheets.xlsx",sheet="UPPS-P")
  uppsp<-uppsp[!is.na(uppsp$ID),]
  uppsp_scoring(uppsp)
}
