% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall_plot}
\alias{waterfall_plot}
\title{waterfall plot}
\usage{
waterfall_plot(d)
}
\arguments{
\item{d}{dataframe with columns id (repeated) and  age}
}
\description{
subj x age plot: line-per-subject dot-per-age ordered by age at first visit
}
\examples{
 d <- data.frame(age=c(10,20,25, 13,14), id=c(100,100,100, 200,200))
 p <- waterfall_plot(d)
 print(p)
}
