% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscorewithinfactor}
\alias{zscorewithinfactor}
\title{zscores within factor for long format/nested data}
\usage{
zscorewithinfactor(fact, score, VERBOSE = T)
}
\arguments{
\item{fact}{is grouping factor}

\item{score}{is variable}

\item{VERBOSE}{flag for printing factor during loop iteration}
}
\description{
zscores within factor for long format/nested data
}
\examples{
  x <- 1:10; f <- sample(c(rep('x',5),rep('y',5)))
  z.vec  <- zscorewithinfactor(f, x)
  z.list <- tapply(x,f,zscore)
}
