% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscorewithinfactor}
\alias{zscorewithinfactor}
\title{zscores within factor for long format/nested data}
\usage{
zscorewithinfactor(fact, score)
}
\arguments{
\item{fact}{is grouping factor}

\item{score}{is variable}
}
\description{
zscores within factor for long format/nested data
}
